// api-client/models/auth-output.ts
/* tslint:disable */
/* eslint-disable */
/**
 * booknest
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { AuthOutputUser } from './auth-output-user';

/**
 * 
 * @export
 * @interface AuthOutput
 */
export interface AuthOutput {
    /**
     * 
     * @type {string}
     * @memberof AuthOutput
     */
    'accessToken': string;
    /**
     * 
     * @type {string}
     * @memberof AuthOutput
     */
    'refreshToken': string;
    /**
     * 
     * @type {string}
     * @memberof AuthOutput
     */
    'type'?: string;
    /**
     * 
     * @type {AuthOutputUser}
     * @memberof AuthOutput
     */
    'user': AuthOutputUser;
}